requirements.txt
pytest==8.3.5
pytest-html==4.1.1
selenium==4.27.1
pyyaml==6.0.3
